adjust api for nextElement() and elements()
The previous implementation of nextElement() ommitted a not-null
constrain on the element type, yet relied on `exists`. The constraint is
important to avoid having to change the code to use a less efficient `is
Element` reified test.

elements() was changed to return an iterable of `Element` rather than
the less useful `Absent|Element`.